+++ src/Reservation.Application/Categories/Queries/GetTop3Category/GetTop3SubCategoryQueryResponse.cs
@@ -3,5 +3,5 @@ namespace Reservation.Application.Categories.Queries.GetTop3Category;
 public record GetTop3SubCategoryQueryResponse
     (
         Guid Id, string Title, string Description,
-        string CoverImagePath, int Average
+        string CoverImagePath, double Average
     ) : IResponse;
